+++ etl/models/models.go
@@ -43,14 +43,14 @@ type Record struct {
 	Hash                []byte
 	RawData             []byte
 	JetID               string
-	PulseNumber         int
+	PulseNumber         int64
 	Order               int
 	Timestamp           int64
 }
 
 type JetDrop struct {
 	JetID          string `gorm:"primary_key;auto_increment:false;default:''"`
-	PulseNumber    int    `gorm:"primary_key;auto_increment:false"`
+	PulseNumber    int64  `gorm:"primary_key;auto_increment:false"`
 	FirstPrevHash  []byte
 	SecondPrevHash []byte
 	Hash           []byte
@@ -60,9 +60,9 @@ type JetDrop struct {
 }
 
 type Pulse struct {
-	PulseNumber     int `gorm:"primary_key;auto_increment:false"`
-	PrevPulseNumber int
-	NextPulseNumber int
+	PulseNumber     int64 `gorm:"primary_key;auto_increment:false"`
+	PrevPulseNumber int64
+	NextPulseNumber int64
 	IsComplete      bool
 	IsSequential    bool
 	Timestamp       int64
